Use %python_provide macros for backward compatibility provides
This allows easier compatibility back and forth once (and if)
https://fedoraproject.org/wiki/Changes/Python_means_Python3
happens. This is not the full implementation of the change,
as the python-unversioned package still stays.

The provides with %{?_isa} are added for backwards compatibility.
In theory, such provides should be added by %python_provide automatically,
but in practice the noarch python-unversioned-command subpackage prevents
it, see https://bugzilla.redhat.com/show_bug.cgi?id=1705656

Also, conditionalize the tkinter provides for Fedora < 31 (nothing uses it).