Allow to disable tkinter support, useful for flatpak builds
This commit resurrects old 04000-modularity-disable-tk.patch patch to
disable tk support, except that this time we're disabling the support
not for modularity, but via a bcond to use when building for flatpak.

The motivation here is to avoid unnecessarily dragging in tcl and tk to
gimp flatpak. No other Fedora flatpak includes python2.7.