Reimport 00361-openssl-3-compat.patch from GitHub
For a while now, patch 361 was incomplete in
https://github.com/fedora-python/cpython/commits/fedora-2.7

This was caused by this dist-git only change:
fd140a8adbfc2317f86bb952df9ff3afdc6e6b80

I've completed the git commit based on the content of the patch here,
pushed it to GitHub, and reimported it here.

As a result, the patch looks slightly different and some of the hashes changed.
But the actual difference should be identical.